+++ lib/votd/base.rb
@@ -152,5 +152,27 @@ module Votd
     def strip_html_tags(text)
       text.gsub(/<\/?[^>]*>/, '')
     end
+
+    # Prepends '...' if first letter is not a capital letter
+    # @param [String] the text to process
+    # @return [String]
+    def clean_verse_start(text)
+      text.sub(/^([a-z])/, '...\1')
+    end
+
+    # Appends '...' if verse ends abruptly
+    # @param [String] the text to process
+    # @return [String]
+    def clean_verse_end(text)
+      case text
+      when /[a-zA=Z]$/         # no ending "."
+        text << '...'
+      when /[,;]$/
+        text.sub!(/,$/, '...') # ends with "," or ";"
+      else
+        text
+      end
+      text
+    end
   end
 end