+++ lib/votd/netbible.rb
@@ -38,10 +38,8 @@ module Votd
 
       # build the text
       text = strip_html_tags(verses.join(" "))
-      # prepend ... if first letter is not a capital letter
-      text.sub!(/^([a-z])/, '...\1')
-      # append ... if last char is anything but a period
-      text.sub!(/[^\.]$/, '...')
+      text = clean_verse_start(text)
+      text = clean_verse_end(text)
 
       @text = text